pyproject.toml
[tool.poetry]
name = "assobot"
version = "0.1.0"
description = ""
authors = ["MIAGE Connection <si@miage.net>"]
license = "MIT License"

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.3"
"discord.py" = "^1.7.3"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
launch-bot = "assobot.app:launch"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
